Migrate from module to library layout.
