+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "entrypoint.py"
-version = "0.1.3"
+version = "0.2.0"
 description = "Decorated functions as entry points."
 authors = ["nekitdev"]
 license = "MIT"
@@ -21,7 +21,7 @@ classifiers = [
     "Typing :: Typed",
 ]
 
-include = ["CHANGELOG.md", "py.typed", "tests"]
+include = ["CHANGELOG.md", "tests", "entrypoint/py.typed"]
 
 [tool.poetry.urls]
 Discord = "https://nekit.dev/discord"
@@ -29,7 +29,7 @@ Funding = "https://patreon.com/nekitdev"
 Tracker = "https://github.com/nekitdev/entrypoint.py/issues"
 
 [[tool.poetry.packages]]
-include = "entrypoint.py"
+include = "entrypoint"
 
 [tool.poetry.dependencies]
 python = ">= 3.7"
@@ -108,7 +108,7 @@ warn_unused_ignores = false  # compatibility
 
 [tool.changelog]
 name = "entrypoint.py"
-version = "0.1.3"
+version = "0.2.0"
 url = "https://github.com/nekitdev/entrypoint.py"
 directory = "changes"
 output = "CHANGELOG.md"
